+++ src/data/index.ts
@@ -11,6 +11,7 @@ export * from './dlc';
 export * from './planets';
 export * from './blocks';
 export * from './functional-blocks';
+export * from './weapon-blocks';
 export * from './generated-blocks';
 export * from './all-blocks';
 export * from './generated-block-costs';